+++ CHANGES.rst
@@ -16,6 +16,7 @@ Other changes:
 
 Bugfix:
  - Fix crash when opening file with external application (#2573)
+ - Fix issue on settings: unable to rename current device if it has no name (#2174)
 
 Translations:
  -